Updated GoogleSheetsLoad to batch load
Google has limitations on how big an upload to Sheets data can be. It is limited to 10 MB. Updated the procedure to batch 1,000,000 characters at a time to work around this limitation. They also have a 5,000,000 cell count limit which we correctly check for now.
Also fixed the GoogleSheetsClear function to resize the document and reset the colors, bold, font, etc.